fix(api/applications): api document publish part 1 tests (BOAS-855)
